refactor: Unify crossruntime Border
