--- src/Uno.UI/UI/Xaml/Controls/Border/Border.skia.cs
@@ -1,88 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Uno.Extensions;
-using System.Linq;
-using System.Drawing;
-using Uno.Disposables;
-using Microsoft.UI.Xaml.Media;
-using Uno.UI;
-
-using View = Microsoft.UI.Xaml.UIElement;
-using Color = System.Drawing.Color;
-using Microsoft.UI.Composition;
-using System.Numerics;
-using Windows.Foundation;
-using Microsoft.UI.Xaml.Shapes;
-using Uno.UI.Xaml.Controls;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class Border
-	{
-		public Border()
-		{
-			BorderRenderer = new BorderLayerRenderer(this);
-		}
-
-		internal BorderLayerRenderer BorderRenderer { get; }
-
-		partial void OnChildChangedPartial(View previousValue, View newValue)
-		{
-			if (previousValue != null)
-			{
-				RemoveChild(previousValue);
-			}
-
-			AddChild(newValue);
-		}
-
-		private void UpdateBorder()
-		{
-			if (Visual != null)
-			{
-				BorderRenderer.Update();
-			}
-		}
-
-		internal override void OnArrangeVisual(Rect rect, Rect? clip)
-		{
-			UpdateBorder();
-
-			base.OnArrangeVisual(rect, clip);
-		}
-
-		private protected override void OnLoaded()
-		{
-			base.OnLoaded();
-			UpdateBorder();
-		}
-
-		partial void OnBorderBrushChangedPartial()
-		{
-			UpdateBorder();
-		}
-
-		partial void OnBorderThicknessChangedPartial(Thickness oldValue, Thickness newValue)
-		{
-			UpdateBorder();
-		}
-
-		partial void OnPaddingChangedPartial(Thickness oldValue, Thickness newValue)
-		{
-			UpdateBorder();
-		}
-
-		partial void OnCornerRadiusUpdatedPartial(CornerRadius oldValue, CornerRadius newValue)
-		{
-			UpdateBorder();
-		}
-
-		protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs args)
-		{
-			base.OnBackgroundChanged(args);
-
-			UpdateBorder();
-		}
-	}
-}
