--- src/Uno.UI/UI/Xaml/Controls/Border/Border.wasm.cs
@@ -1,57 +1,0 @@
-﻿namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class Border
-	{
-		public Border()
-		{
-			this.SizeChanged += (_, _) => UpdateBorder();
-		}
-
-		partial void OnChildChangedPartial(UIElement previousValue, UIElement newValue)
-		{
-			if (previousValue != null)
-			{
-				RemoveChild(previousValue);
-			}
-
-			AddChild(newValue);
-		}
-
-		private void UpdateBorder()
-		{
-			SetBorder(BorderThickness, BorderBrush, CornerRadius);
-		}
-
-		private protected override void OnLoaded()
-		{
-			base.OnLoaded();
-			UpdateBorder();
-		}
-
-		partial void OnBorderBrushChangedPartial()
-		{
-			UpdateBorder();
-		}
-
-		partial void OnBorderThicknessChangedPartial(Thickness oldValue, Thickness newValue)
-		{
-			UpdateBorder();
-		}
-
-		partial void OnPaddingChangedPartial(Thickness oldValue, Thickness newValue)
-		{
-			UpdateBorder();
-		}
-
-		partial void OnCornerRadiusUpdatedPartial(CornerRadius oldValue, CornerRadius newValue)
-		{
-			UpdateBorder();
-		}
-
-		protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs e)
-		{
-			base.OnBackgroundChanged(e);
-			UpdateHitTest();
-		}
-	}
-}
